Feature/social update UI

+++ app/src/main/java/com/example/tiktok_clone/features/social/ui/components/FormatCount.kt
@@ -1,7 +1,7 @@
 package com.example.tiktok_clone.features.social.ui.components
 
 // định dạng đếm số
-fun formatCount(count: Int): String {
+fun formatCount(count: Long): String {
     return when {
         count >= 1_000_000 -> "%.1fM".format(count / 1_000_000.0)
         count >= 1_000 -> "%.1fk".format(count / 1_000.0)
